(py-process-filter): Deleted this function.  In order to fix interactions with newer Emacsen, I've rewritten the way all the process filters work in the *Python* buffer.  We use more of the comint infrastructure, specifically the default process filter.  This means that scrolling is now handled by the default comint variables including comint-scroll-to-bottom-on-output.  Note that this is somewhat experimental change!
(py-comint-output-filter-function): Moved to here from the obsolete
py-process-filter function, the logic to pop and exec the next queued
file waiting to be executed.

(py-execute-file): Don't bind comint-scroll-to-bottom-on-output to t,
and save the excursion when inserting the "working on" message.  This
lets the standard comint scrolling variables as set by the user,
continue to work.

(python-mode, py-shell, py-describe-mode): Remove description of
py-scroll-process-buffer.  Also in py-shell, make
comint-output-filter-functions buffer-local, and add
py-comint-output-filter-function to this hook (instead of setting the
process filter).

(py-scroll-process-buffer): Deleted this variable.  See comint
variables including comint-scroll-to-bottom-on-output.

(py-execute-region): When exec files are being queued, push the next
temp file on the end of the list.

(py-submit-bug-report): Removed reporting of py-scroll-process-buffer.


git-svn-id: http://svn.python.org/projects/python/trunk@11521 6015fed2-1504-0410-9fe1-9d1591cc4771