Jeremy writes:
I found the following patch helpful in tracking down a bug in some
code.  I had appended time, the module, instead of time.time().  Not
sure if it is generally true that printing the repr of the object is
good, but I expect that most unpicklable things will have fairly
information and concise reprs (like files or sockets or modules).


git-svn-id: http://svn.python.org/projects/python/trunk@13920 6015fed2-1504-0410-9fe1-9d1591cc4771