Patch #101121, by Ka-Ping Yee: cosmetic cleanup of cgi.py, using my style conventions.  (Ping has checkin privileges but apparently ignores them at the moment.)
Ping improves a few doc strings and fixes style violations like foo ( bar ).

An addition of my own: rearrange the printing of various items in
test() so that the (long) environment comes at the end.  This avoids
having to scroll if you want to see the current directory or command
line arguments.


git-svn-id: http://svn.python.org/projects/python/trunk@17532 6015fed2-1504-0410-9fe1-9d1591cc4771