Change RuntimeError to SGMLParseError, which subclasses RuntimeError for backward compatibility.
Add support for SGML declaration syntax (<!....>) to some reasonable
degree.  This does not support everything allowed in SGML, but should
work with "real" HTML (internal subset in a DOCTYPE is not handled).
The content of the declaration is passed to the .handle_decl() method,
which can be overridden by subclasses.


git-svn-id: http://svn.python.org/projects/python/trunk@19951 6015fed2-1504-0410-9fe1-9d1591cc4771