This patch turns the Python API mismatch notice into a standard Python warning which can be catched by means of the Python warning framework.
It also adds two new APIs which hopefully make it easier for Python
to switch to buffer overflow safe [v]snprintf() APIs for error
reporting et al. The two new APIs are PyOS_snprintf() and
PyOS_vsnprintf() and work just like the standard ones in many
C libs. On platforms which have snprintf(), the native APIs are used,
on all other an emulation with snprintf() tries to do its best.


git-svn-id: http://svn.python.org/projects/python/trunk@22105 6015fed2-1504-0410-9fe1-9d1591cc4771