Patch #553678: Cygwin Makefile.pre.in vestige patch
This patch removes a vestige part of the Cygwin make rules
that didn't quite make it over during the flattening of the
Makefiles. In its current form, it creates a def file but
incorrectly calls it libpython$(VERSION).dll.a which
immediately gets overwritten by the next command.
Obviously, this is useless. It appears, it was useless
in the old nested Makefile structure too. :,)


git-svn-id: http://svn.python.org/projects/python/trunk@26820 6015fed2-1504-0410-9fe1-9d1591cc4771