Fix for SF bug #432621: httplib: multiple Set-Cookie headers
If multiple header fields with the same name occur, they are combined
according to the rules in RFC 2616 sec 4.2:

Appending each subsequent field-value to the first, each separated by
a comma. The order in which header fields with the same field-name are
received is significant to the interpretation of the combined field
value.


git-svn-id: http://svn.python.org/projects/python/trunk@27491 6015fed2-1504-0410-9fe1-9d1591cc4771