Fix for SF bug 601077 by Zack Weinberg.
The new execvpe code would sometimes do the wrong thing when a
non-executable file existed earlier in the path and an executable file
of the same name existed later in the path.  This patch restores the
proper behavior (which is to execute the second file).  When only a
non-executable file exists, the correct error is still reported.


git-svn-id: http://svn.python.org/projects/python/trunk@28531 6015fed2-1504-0410-9fe1-9d1591cc4771