On Mac OS X pwd.getpwall() might return the tuples ('pgsql', '*', 252, []) and ('postgres', '*', 252, ['skip']), but pwd.getgrgid(252) might return ('pgsql', '', 252, ['skip']).
Drop the test that tried to find a tuple similar to the one
returned from pwd.getgrgid() among those for the same gid returned
by pwd.getgrall(), as the only working definition of 'similar' seems
to be 'has the same gid'. This check can be done more directly.

This should fix SF bug #732783.


git-svn-id: http://svn.python.org/projects/python/trunk@32633 6015fed2-1504-0410-9fe1-9d1591cc4771