Remove many blanket try/except clauses. SF bug [ 751276 ] cPickle doesn't raise error, pickle does (recursiondepth)
Most of the calls to PyErr_Clear() were intended to catch & clear an
attribute error and try something different.  Guard all those cases
with a PyErr_ExceptionMatches() and fail if some other error
occurred.  The other error is likely a bug in the user code.

This is basically the C equivalent of changing "except:" to
"except AttributeError:"


git-svn-id: http://svn.python.org/projects/python/trunk@33162 6015fed2-1504-0410-9fe1-9d1591cc4771