Add PyThreadState_SetAsyncExc(long, PyObject *).
A new API (only accessible from C) to interrupt a thread by sending it
an exception.  This is not always effective, but might help some people.
Requested by Just van Rossum and Alex Martelli.  It is intentional
that you have to write your own C extension to call it from Python.

Docs will have to wait.


git-svn-id: http://svn.python.org/projects/python/trunk@33306 6015fed2-1504-0410-9fe1-9d1591cc4771