Fix the most recent change to the invocation of the mkhowto script so that it works for all targets.
The issue here is that there are two different levels in the directory
tree at which we execute mkhowto, so we can't define it just once
using a relative path (at least not with the current implementation
and Makefile structure).  We use the GNUish $(shell) function here to
work around that restriction by identifying mkhowto using an absolute
path.


git-svn-id: http://svn.python.org/projects/python/trunk@34267 6015fed2-1504-0410-9fe1-9d1591cc4771