PyErr_ProgramText():  Grrrrrr.
In a Windows debug build, trying to open a file using
an empty string as the name causes assertion death
inside MS's C runtime code.  We probably need to worm
around that in many places.  I'm worming around it here
to stop the new test_with.py from assert-dying in the
Windows debug build (it calls compile() with an empty
string for "the file name", which indirectly leads to
C-level code in Python trying to fopen("", "r")).


git-svn-id: http://svn.python.org/projects/python/trunk@42631 6015fed2-1504-0410-9fe1-9d1591cc4771