Do not copy free variables to locals in class namespaces.
Fixes bug 1569356, but at the cost of a minor incompatibility in
locals().  Add test that verifies that the class namespace is not
polluted.  Also clarify the behavior in the library docs.

Along the way, cleaned up the dict_to_map and map_to_dict
implementations and added some comments that explain what they do.



git-svn-id: http://svn.python.org/projects/python/trunk@53954 6015fed2-1504-0410-9fe1-9d1591cc4771