Add 'U'/'U#' format characters to Py_BuildValue (and thus to PyObject_CallFunction()) that take a char * (and a size in the case of 'U#') and create a unicode object out of it.
Add functions PyUnicode_FromFormat() and PyUnicode_FromFormatV()
that work similar to PyString_FromFormat(), but create a unicode
object (also a %U format character has been added, that takes
a PyObject *, which must point to a unicode object).

Change the encoding and reason attributes of UnicodeEncodeError,
UnicodeDecodeError and UnicodeTranslateError to be unicode
objects.


git-svn-id: http://svn.python.org/projects/python/branches/py3k-struni@55433 6015fed2-1504-0410-9fe1-9d1591cc4771