Rip out the file object's implementation. Fixed test_import.py while I was at it.
However, there's still a problem in import.c -- get_file() can leak a
FILE struct (not a file descriptor though).  I'm not sure how to fix
this; closing the FILE* closes the file descriptor, and that's the
wrong thing to do when there's still a Python file object keeping the
file descriptor open.  I also would rather not mess with dup(), as it
won't port to Windows.



git-svn-id: http://svn.python.org/projects/python/branches/py3k-struni@55943 6015fed2-1504-0410-9fe1-9d1591cc4771