Now in find, rfind, index, and rindex, you can use None as defaults, as usual with slicing (both with str and unicode strings).  This fixes issue 1259.
For str only the stringobject.c file was modified.  But for unicode,
I needed to repeat in the four functions a lot of code, so created 
a new function that does part of the job for them (and placed it in 
find.h, following a suggestion of Barry).

Also added tests for this behaviour.


git-svn-id: http://svn.python.org/projects/python/trunk@59020 6015fed2-1504-0410-9fe1-9d1591cc4771