Silence the DeprecationWarning of rfc822 triggered by its importation in mimetools.
This has an unfortunate side-effect of potentially not letting any warning
about rfc822's deprecation be seen by user-visible code if rfc822 is not
imported before mimetools. This is because modules are cached in sys.modules
and thus do not have their deprecation triggered more than once. But this
silencing would have happened by other code that silences the use of mimetools
or rfc822 anyway in the stdlib or user code, and thus seems justified to be
done here.


git-svn-id: http://svn.python.org/projects/python/trunk@65665 6015fed2-1504-0410-9fe1-9d1591cc4771